sel4utils: properly cache ept pages (#43)
A kernel bug previously always cached EPT page regardless of the
cacheable flag. When the bug was fixed, this function would always map
EPT pages as uncacheable, due to the attributes difference between an
EPT map and a normal page map. This commit checks if the function is
mapping an EPT and provides the proper cache attribute to the kernel.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>
GitOrigin-RevId: dd02025f89e9813fd39ff875d3c81056ccfc0bbc